refactor: replace ModelConfigDialog with a shared component and update related imports for better organization and clarity
